+++ src/components/Tabs.vue
@@ -46,6 +46,8 @@
         data: () => ({
             tabs: [],
             activeTabHash: '',
+            activeTabIndex: 0,
+            lastActiveTabHash: '',
         }),
 
         computed: {
@@ -104,6 +106,11 @@
                     return;
                 }
 
+                if (this.lastActiveTabHash === selectedTab.hash) {
+                    this.$emit('clicked', { tab: selectedTab });
+                    return;
+                }
+
                 this.tabs.forEach(tab => {
                     tab.isActive = (tab.hash === selectedTab.hash);
                 });
@@ -111,6 +118,9 @@
                 this.$emit('changed', { tab: selectedTab });
 
                 this.activeTabHash = selectedTab.hash;
+                this.activeTabIndex = this.getTabIndex(selectedTabHash);
+
+                this.lastActiveTabHash = this.activeTabHash = selectedTab.hash;
 
                 expiringStorage.set(this.storageKey, selectedTab.hash, this.cacheLifetime);
             },
@@ -138,6 +148,30 @@
                         return true;
                     });
                 }
+            },
+            
+            getTabIndex(hash){
+            	const tab = this.findTab(hash);
+            	
+            	return this.tabs.indexOf(tab);
+            },
+            
+			getTabHash(index){
+            	const tab = this.tabs.find(tab => this.tabs.indexOf(tab) === index);
+            	
+            	if (!tab) {
+					return;
+                }
+                
+                return tab.hash;
+			},
+            
+            getActiveTab(){
+            	return this.findTab(this.activeTabHash);
+            },
+            
+			getActiveTabIndex() {
+            	return this.getTabIndex(this.activeTabHash);
             },
         },
     };